requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

langchain==0.0.333
gpt4all==2.0.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
azure-functions
azure-data-tables
azure-storage-queue
